requirements.txt
python-telegram-bot>=20.0
openai>=1.0.0
pydub>=0.25.1
requests>=2.32
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
alembic>=1.12.0
